requirements.txt
python-decouple
discord
pillow
praw